+++ crates/rtlola-frontend/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.4.3] - ?
 
+### Changed
+- lowering of cast from function to convert
+
 ### Fixed
 - Sqrt and other functions expected value of size 0